Updated snake design
Updated both the draw_snake and draw_ghostly_snake functions. Snake head is now bigger, has eyes, and on the correct end of the body. Snake body is composed of ellipses currently but will be updated. Ghost snake is the same design but white with blue eyes to better resemble that of a ghost.